requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4"
anyio==4.11.0 ; python_version >= "3.11" and python_version < "4"
beautifulsoup4==4.14.2 ; python_version >= "3.11" and python_version < "4"
cachetools==6.2.1 ; python_version >= "3.11" and python_version < "4"
certifi==2025.10.5 ; python_version >= "3.11" and python_version < "4"
charset-normalizer==3.4.4 ; python_version >= "3.11" and python_version < "4"
click==8.3.0 ; python_version >= "3.11" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4" and platform_system == "Windows"
dnspython==2.8.0 ; python_version >= "3.11" and python_version < "4"
extruct==0.18.0 ; python_version >= "3.11" and python_version < "4"
fastapi==0.119.1 ; python_version >= "3.11" and python_version < "4"
google-auth==2.41.1 ; python_version >= "3.11" and python_version < "4"
google-genai==1.45.0 ; python_version >= "3.11" and python_version < "4"
h11==0.16.0 ; python_version >= "3.11" and python_version < "4"
html-text==0.7.1 ; python_version >= "3.11" and python_version < "4"
html5lib==1.1 ; python_version >= "3.11" and python_version < "4"
httpcore==1.0.9 ; python_version >= "3.11" and python_version < "4"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4"
idna==3.11 ; python_version >= "3.11" and python_version < "4"
isodate==0.7.2 ; python_version >= "3.11" and python_version < "4"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4"
jstyleson==0.0.2 ; python_version >= "3.11" and python_version < "4"
lxml-html-clean==0.4.3 ; python_version >= "3.11" and python_version < "4"
lxml==6.0.2 ; python_version >= "3.11" and python_version < "4"
markupsafe==3.0.3 ; python_version >= "3.11" and python_version < "4"
mf2py==2.0.1 ; python_version >= "3.11" and python_version < "4"
notion-client==2.6.0 ; python_version >= "3.11" and python_version < "4"
pyasn1-modules==0.4.2 ; python_version >= "3.11" and python_version < "4"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4"
pydantic-core==2.41.4 ; python_version >= "3.11" and python_version < "4"
pydantic==2.12.3 ; python_version >= "3.11" and python_version < "4"
pymongo==4.15.3 ; python_version >= "3.11" and python_version < "4"
pyparsing==3.2.5 ; python_version >= "3.11" and python_version < "4"
pyrdfa3==3.6.4 ; python_version >= "3.11" and python_version < "4"
rdflib==7.2.1 ; python_version >= "3.11" and python_version < "4"
recipe-scrapers==15.9.0 ; python_version >= "3.11" and python_version < "4"
requests==2.32.5 ; python_version >= "3.11" and python_version < "4"
rsa==4.2 ; python_version >= "3.11" and python_version < "4"
six==1.17.0 ; python_version >= "3.11" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4"
soupsieve==2.8 ; python_version >= "3.11" and python_version < "4"
starlette==0.48.0 ; python_version >= "3.11" and python_version < "4"
tenacity==9.1.2 ; python_version >= "3.11" and python_version < "4"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_version < "4"
typing-inspection==0.4.2 ; python_version >= "3.11" and python_version < "4"
urllib3==2.5.0 ; python_version >= "3.11" and python_version < "4"
uvicorn==0.38.0 ; python_version >= "3.11" and python_version < "4"
w3lib==2.3.1 ; python_version >= "3.11" and python_version < "4"
webencodings==0.5.1 ; python_version >= "3.11" and python_version < "4"
websockets==15.0.1 ; python_version >= "3.11" and python_version < "4"
